sandbox/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
google-genai>=1.0.0
